fix(g-base): canvas matrix should affect totalMatrix of its children, close #402
